requirements.txt
asgiref==3.2.4
astroid==2.4.1
autopep8==1.5.4
awsebcli==3.18.1
boto3==1.14.56
botocore==1.17.56
cachetools==4.1.1
cement==2.8.2
certifi==2019.11.28
cffi==1.14.2
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.4
django-address==0.2.2
django-allauth==0.41.0
django-autoslug==1.9.7
django-cors-headers==3.3.0
django-countries==6.1.2
django-decouple==2.1
django-environ==0.4.5
django-filter==2.3.0
django-heroku==0.3.1
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-sendgrid-v5==0.8.1
django-storages==1.7.1
django-utils-six==2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
docutils==0.15.2
future==0.16.0
google-api-core==1.22.1
google-auth==1.21.0
google-cloud-core==0.28.1
google-cloud-storage==1.10.0
google-crc32c==1.0.0
google-resumable-media==1.0.0
googleapis-common-protos==1.52.0
gunicorn==20.0.4
idna==2.7
importlib-metadata==1.7.0
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
openapi-codec==1.3.2
pathspec==0.5.9
Pillow==7.0.0
portalocker==2.0.0
protobuf==3.13.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pylint==2.5.2
python-dateutil==2.8.0
python-http-client==3.3.0
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3.1
requests==2.20.1
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
semantic-version==2.5.0
sendgrid==6.4.6
simplejson==3.17.0
six==1.11.0
sqlparse==0.3.1
starkbank-ecdsa==1.0.0
termcolor==1.1.0
toml==0.10.1
typed-ast==1.4.1
Unipath==1.1
uritemplate==3.0.1
urllib3==1.24.3
wcwidth==0.1.9
whitenoise==5.1.0
wrapt==1.12.1
zipp==3.1.0
